Renamed "start-node" to "source-picker"
e.g. `<umb-property-editor-ui-tree-picker-source-picker>`
