+++ src/packages/core/property-editor/uis/tree-picker/config/source-picker/manifests.ts
@@ -2,11 +2,11 @@ import type { ManifestPropertyEditorUi } from '@umbraco-cms/backoffice/extension
 
 export const manifest: ManifestPropertyEditorUi = {
 	type: 'propertyEditorUi',
-	alias: 'Umb.PropertyEditorUi.TreePicker.StartNode',
-	name: 'Tree Picker Start Node Property Editor UI',
-	js: () => import('./property-editor-ui-tree-picker-start-node.element.js'),
+	alias: 'Umb.PropertyEditorUi.TreePicker.SourcePicker',
+	name: 'Tree Picker Source Picker Property Editor UI',
+	js: () => import('./property-editor-ui-tree-picker-source-picker.element.js'),
 	meta: {
-		label: 'Tree Picker Start Node',
+		label: 'Tree Picker Source Picker',
 		icon: 'icon-page-add',
 		group: 'pickers',
 		propertyEditorSchemaAlias: '',
