+++ src/packages/core/property-editor/uis/tree-picker/config/source-picker/property-editor-ui-tree-picker-source-picker.test.ts
@@ -1,0 +1,23 @@
+import { expect, fixture, html } from '@open-wc/testing';
+import { UmbPropertyEditorUITreePickerSourcePickerElement } from './property-editor-ui-tree-picker-source-picker.element.js';
+import { defaultA11yConfig } from '@umbraco-cms/internal/test-utils';
+
+describe('UmbPropertyEditorUITreePickerSourcePickerElement', () => {
+	let element: UmbPropertyEditorUITreePickerSourcePickerElement;
+
+	beforeEach(async () => {
+		element = await fixture(html`
+			<umb-property-editor-ui-tree-picker-source-picker></umb-property-editor-ui-tree-picker-source-picker>
+		`);
+	});
+
+	it('is defined with its own instance', () => {
+		expect(element).to.be.instanceOf(UmbPropertyEditorUITreePickerSourcePickerElement);
+	});
+
+	if ((window as any).__UMBRACO_TEST_RUN_A11Y_TEST) {
+		it('passes the a11y audit', async () => {
+			await expect(element).shadowDom.to.be.accessible(defaultA11yConfig);
+		});
+	}
+});
